+++ src/main/java/com/develhope/Java27_progetto3_team2/courier/Courier.java
@@ -1,4 +1,4 @@
-package com.develhope.Java27_progetto3_team2.entity;
+package com.develhope.Java27_progetto3_team2.courier;
 
 import jakarta.persistence.*;
 import lombok.*;
