+++ src/main/java/com/develhope/Java27_progetto3_team2/order/OrderRepository.java
@@ -1,6 +1,5 @@
-package com.develhope.Java27_progetto3_team2.repository;
+package com.develhope.Java27_progetto3_team2.order;
 
-import com.develhope.Java27_progetto3_team2.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
