+++ src/main/java/com/develhope/Java27_progetto3_team2/order/OrderService.java
@@ -1,9 +1,5 @@
-package com.develhope.Java27_progetto3_team2.service;
+package com.develhope.Java27_progetto3_team2.order;
 
-import com.develhope.Java27_progetto3_team2.entity.Order;
-import com.develhope.Java27_progetto3_team2.mapper.OrderMapper;
-import com.develhope.Java27_progetto3_team2.model.OrderDTO;
-import com.develhope.Java27_progetto3_team2.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
